Migrados endpoints de electivas a VSF
