+++ api/_helpers/jwt.js
@@ -11,10 +11,8 @@ function jwt() {
       "/users/authenticate",
       "/backend",
       "/cards/getAll",
-      "/api/electivas/get",
       "/api/deportes/get",
       "/api/deportes/getAll",
-      "/electivas",
       "/deportes",
       "/calendario",
     ],
